[LF] Drop kv-support
