--- daml-lf/kv-support/BUILD.bazel
@@ -1,61 +1,0 @@
-# Copyright (c) 2023 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-load(
-    "//bazel_tools:scala.bzl",
-    "da_scala_library",
-    "da_scala_test_suite",
-    "lf_scalacopts",
-    "lf_scalacopts_stricter",
-)
-
-da_scala_library(
-    name = "kv-support",
-    srcs = glob(["src/main/**/*.scala"]),
-    scala_deps = [
-        "@maven//:org_scalaz_scalaz_core",
-    ],
-    scalacopts = lf_scalacopts_stricter,
-    tags = ["maven_coordinates=com.daml:daml-lf-kv-support:__VERSION__"],
-    deps = [
-        "//daml-lf/archive:daml_lf_1.dev_archive_proto_java",
-        "//daml-lf/archive:daml_lf_archive_reader",
-        "//daml-lf/data",
-        "//daml-lf/language",
-        "//daml-lf/transaction",
-        "//daml-lf/transaction:transaction_proto_java",
-        "//daml-lf/transaction:value_proto_java",
-        "//libs-scala/safe-proto",
-        "@maven//:com_google_protobuf_protobuf_java",
-    ],
-)
-
-da_scala_test_suite(
-    name = "test",
-    srcs = glob(["src/test/**/*.scala"]),
-    scala_deps = [
-        "@maven//:org_scalaz_scalaz_core",
-        "@maven//:org_scalacheck_scalacheck",
-        "@maven//:org_scalatest_scalatest_core",
-        "@maven//:org_scalatest_scalatest_matchers_core",
-        "@maven//:org_scalatest_scalatest_shouldmatchers",
-        "@maven//:org_scalatest_scalatest_wordspec",
-        "@maven//:org_scalatestplus_scalacheck_1_15",
-    ],
-    scalacopts = lf_scalacopts,
-    deps = [
-        ":kv-support",
-        "//daml-lf/archive:daml_lf_1.dev_archive_proto_java",
-        "//daml-lf/archive:daml_lf_archive_reader",
-        "//daml-lf/data",
-        "//daml-lf/encoder",
-        "//daml-lf/language",
-        "//daml-lf/parser",
-        "//daml-lf/transaction",
-        "//daml-lf/transaction:transaction_proto_java",
-        "//daml-lf/transaction:value_proto_java",
-        "//daml-lf/transaction-test-lib",
-        "@maven//:com_google_protobuf_protobuf_java",
-        "@maven//:org_scalatest_scalatest_compatible",
-    ],
-)
