--- daml-lf/kv-support/README.md
@@ -1,7 +1,0 @@
-The KV Daml-LF library
-======================
-
-This package contains LF utilities specifically tailored for the KV
-integration kit. This library does not define a stable public API, and
-should not be used outside of the KV integration kit.
-Breaking changes to this library will be made without notice.
