--- daml-lf/kv-support/src/main/scala/com/daml/lf/kv/archives/RawArchive.scala
@@ -1,9 +1,0 @@
-// Copyright (c) 2023 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.daml.lf.kv.archives
-
-import com.google.protobuf.ByteString
-
-/** Stores [[com.daml.daml_lf_dev.DamlLf.Archive]] as a [[ByteString]]. */
-case class RawArchive(byteString: ByteString) extends AnyVal
