--- daml-lf/kv-support/src/main/scala/com/daml/lf/kv/contracts/RawContractInstance.scala
@@ -1,9 +1,0 @@
-// Copyright (c) 2023 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.daml.lf.kv.contracts
-
-import com.google.protobuf.ByteString
-
-/** Stores [[com.daml.lf.transaction.TransactionOuterClass.ContractInstance]] as a [[ByteString]]. */
-case class RawContractInstance(byteString: ByteString) extends AnyVal
