+++ daml-lf/transaction/src/main/scala/com/digitalasset/daml/lf/transaction/TransactionCoder.scala
@@ -837,94 +837,6 @@ object TransactionCoder {
     else
       decodeVersion(node.getVersion)
 
-  /** Action node information for a serialized transaction node. Used to compute
-    * informees when deserialization is too costly.
-    * This method is not supported for transaction version <5 (as NodeInfo does not support it).
-    * We're not using e.g. "implicit class" in order to keep the decoding errors explicit.
-    * NOTE(JM): Currently used only externally, but kept here to keep in sync
-    * with the implementation.
-    * Note that this can only be applied to action nodes and will return Left on
-    * rollback nodes.
-    */
-  def protoActionNodeInfo(
-      txVersion: TransactionVersion,
-      protoNode: TransactionOuterClass.Node,
-  ): Either[DecodeError, ActionNodeInfo] =
-    protoNode.getNodeTypeCase match {
-      case NodeTypeCase.ROLLBACK =>
-        Left(
-          DecodeError(
-            "protoActionNodeInfo only supports action nodes but was applied to a rollback node"
-          )
-        )
-      case NodeTypeCase.CREATE =>
-        val protoCreate = protoNode.getCreate
-        for {
-          signatories_ <- toPartySet(protoCreate.getSignatoriesList)
-          stakeholders_ <- toPartySet(protoCreate.getStakeholdersList)
-        } yield {
-          new ActionNodeInfo.Create {
-            def signatories = signatories_
-            def stakeholders = stakeholders_
-          }
-        }
-      case NodeTypeCase.FETCH =>
-        val protoFetch = protoNode.getFetch
-        for {
-          actingParties_ <- toPartySet(protoFetch.getActorsList)
-          stakeholders_ <- toPartySet(protoFetch.getStakeholdersList)
-          signatories_ <- toPartySet(protoFetch.getSignatoriesList)
-        } yield {
-          new ActionNodeInfo.Fetch {
-            def signatories = signatories_
-            def stakeholders = stakeholders_
-            def actingParties = actingParties_
-          }
-        }
-
-      case NodeTypeCase.EXERCISE =>
-        val protoExe = protoNode.getExercise
-        for {
-          actingParties_ <- toPartySet(protoExe.getActorsList)
-          signatories_ <- toPartySet(protoExe.getSignatoriesList)
-          stakeholders_ <- toPartySet(protoExe.getStakeholdersList)
-          choiceObservers_ <-
-            if (txVersion < TransactionVersion.minChoiceObservers)
-              Right(Set.empty[Party])
-            else
-              toPartySet(protoExe.getObserversList)
-          choiceAuthorizers_ <-
-            if (txVersion < TransactionVersion.minChoiceAuthorizers) {
-              Right(None)
-            } else
-              toPartySet(protoExe.getAuthorizersList).map(choiceAuthorizersList =>
-                Option.when(choiceAuthorizersList.nonEmpty)(choiceAuthorizersList)
-              )
-        } yield {
-          new ActionNodeInfo.Exercise {
-            def signatories = signatories_
-            def stakeholders = stakeholders_
-            def actingParties = actingParties_
-            def choiceObservers = choiceObservers_
-            def choiceAuthorizers = choiceAuthorizers_
-            def consuming = protoExe.getConsuming
-          }
-        }
-
-      case NodeTypeCase.LOOKUP_BY_KEY =>
-        val protoLookupByKey = protoNode.getLookupByKey
-        for {
-          maintainers <- toPartySet(protoLookupByKey.getKeyWithMaintainers.getMaintainersList)
-        } yield {
-          new ActionNodeInfo.LookupByKey {
-            def hasResult = protoLookupByKey.hasContractIdStruct
-            def keyMaintainers = maintainers
-          }
-        }
-
-      case NodeTypeCase.NODETYPE_NOT_SET => Left(DecodeError("Unset Node type"))
-    }
-
   private[this] def keyHash(
       nodeVersion: TransactionVersion,
       rawTmplId: ValueOuterClass.Identifier,
